Added 2.2.5

+++ uikit/src/main/java/com/sendbird/uikit/utils/UIKitPrefs.java
@@ -17,11 +17,13 @@ final public class UIKitPrefs {
     }
 
     public static void clearAll() {
+        if (preferences == null) return;
         SharedPreferences.Editor editor = preferences.edit();
         editor.clear().apply();
     }
 
     public static void remove(String key ) {
+        if (preferences == null) return;
         if (preferences.contains(key)) {
             SharedPreferences.Editor editor = preferences.edit();
             editor.remove(key).apply();
@@ -29,6 +31,7 @@ final public class UIKitPrefs {
     }
 
     public static void putString(String key, String value) {
+        if (preferences == null) return;
         SharedPreferences.Editor editor = preferences.edit();
         editor.putString(key, value).apply();
     }
@@ -38,10 +41,12 @@ final public class UIKitPrefs {
     }
 
     public static String getString(String key, String defValue) {
+        if (preferences == null) return defValue;
         return preferences.getString(key, defValue);
     }
 
     public static void putLong(String key, long value) {
+        if (preferences == null) return;
         SharedPreferences.Editor editor = preferences.edit();
         editor.putLong(key, value).apply();
     }
@@ -51,10 +56,12 @@ final public class UIKitPrefs {
     }
 
     public static long getLong(String key, long defValue) {
+        if (preferences == null) return defValue;
         return preferences.getLong(key, defValue);
     }
 
     public static void putInt(String key, int value) {
+        if (preferences == null) return;
         SharedPreferences.Editor editor = preferences.edit();
         editor.putInt(key, value).apply();
     }
@@ -64,10 +71,12 @@ final public class UIKitPrefs {
     }
 
     public static int getInt(String key, int defValue) {
+        if (preferences == null) return defValue;
         return preferences.getInt(key, defValue);
     }
 
     public static void putBoolean(String key, boolean value) {
+        if (preferences == null) return;
         SharedPreferences.Editor editor = preferences.edit();
         editor.putBoolean(key, value).apply();
     }
@@ -77,6 +86,7 @@ final public class UIKitPrefs {
     }
 
     public static boolean getBoolean(String key, boolean defValue) {
+        if (preferences == null) return defValue;
         return preferences.getBoolean(key, defValue);
     }
 }
